requirements.txt
django
psycopg2-binary
gunicorn
django-two-factor-auth
django-environ
python-dotenv
django-crispy-forms
crispy-bootstrap5
django-cryptography
django-encrypted-model-fields
stripe
django-csp
sendgrid
celery
redis
langchain
langchain-openai
langgraph
langchain-core
pydantic
django-celery-beat
requests_oauthlib
django-allauth
pyjwt
bs4